backend/requirements.txt
# ProofCart Backend Requirements

# Django & DRF
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
django-filter==23.5
drf-yasg==1.21.11

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Blockchain Integration
solana==0.30.2
solders==0.18.1
anchorpy==0.18.0
ic-py==1.0.1

# Environment & Config
python-decouple==3.8
python-dotenv==1.0.0

# Utilities
Pillow==10.4.0
requests==2.31.0
qrcode==8.2

# IPFS/NFT Storage
ipfshttpclient==0.8.0a2

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Development
django-debug-toolbar==4.2.0
